Get Best from cluster
This program takes as input a file containing a set of clusters of proteins (1 cluster per line and 1 PDB ID per column) and a file containing the PDB ID's and the resolution of each structure and retrieves the member of each cluster that has the highest resolution. This can be implemented from the bash console.